ENH: added fitmultbsid library
Added a library for the FitMultBSID structures and supporting functions.  Updated make_fov to use this new library.
