+++ codebase/superdarn/src.bin/tk/tool/make_fov.1.0/eval_groundscatter.c
@@ -15,7 +15,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#include "multbsid.h"
+#include "fitmultbsid.h"
 
 /**
  * @brief Update the groundscatter flag 'gflg' for points already flagged as GS
