+++ codebase/superdarn/src.bin/tk/tool/make_fov.1.0/makefile
@@ -22,14 +22,12 @@ SRC =	make_fov.c \
 	linear_regression.c \
 	load_fit_update_fov.c \
 	load_radar_site.c \
-	multbsid.c \
 	scan_utils.c \
 	select_alt_groups.c \
 	stat_utils.c \
 	test_ut_fov_struct.c \
 	update_beam.c \
-	update_scan.c \
-	write_scan.c
+	update_scan.c
 
 OBJS =	$(SRC:.c=.o)
 
@@ -43,7 +41,7 @@ LIBS = -lchannel.1 -lfreqband.1 -loldfit.1 -lgtabw.1 -loldgtabw.1 -lgtable.1 \
 	-lrpos.1 -lfilter.1 -lcfit.1 -lfit.1 -lrscan.1 -lradar.1 -ldmap.1 \
 	-lopt.1 -lrtime.1 -lrcnv.1 -laacgm.1 -ligrf.1 -laacgm_v2.1 -ligrf_v2.1 \
 	-lastalg.1 -lfitacf.1 -lfitacf.3.0 -lmpfit.1.2 -lelevation.1.0 \
-	-lpropagation.1.0 -lsort.1.0
+	-lpropagation.1.0 -lsort.1.0 -lfitmultbsid.1.0
 
 SLIB = -lm -lz
