--- codebase/superdarn/src.bin/tk/tool/make_fov.1.0/multbsid.c
@@ -1,103 +1,0 @@
-/* multbisd.c
-   ===========
-   Author A.G.Burrell
-*/
-
-/*
- LICENSE AND DISCLAIMER
-
- Copyright (c) 2019 NRL
-
- This file is part of the Radar Software Toolkit (RST).
-
- RST is free software: you can redistribute it and/or modify
- it under the terms of the GNU Lesser General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- any later version.
-
- RST is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU Lesser General Public License for more details.
-
- You should have received a copy of the GNU Lesser General Public License
- along with RST.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <sys/types.h>
-
-#include "multbsid.h"
-
-
-struct MultFitBSID *MultFitBSIDMake()
-{
-  struct MultFitBSID *ptr;
-
-  ptr = (struct MultFitBSID *)malloc(sizeof(struct MultFitBSID));
-  memset(ptr, 0, sizeof(struct MultFitBSID));
-  ptr->num_scans = 0;
-  ptr->scan_ptr  = (struct FitBSIDScan *)(NULL);
-  ptr->last_ptr  = ptr->scan_ptr;
-
-  return ptr;
-}
-
-void MultFitBSIDFree(struct MultFitBSID *ptr)
-{
-  if(MultFitBSIDReset(ptr) == 0) return;
-  free(ptr);
-  return;
-}
-
-int MultFitBSIDReset(struct MultFitBSID *ptr)
-{
-  int n;
-  struct FitBSIDScan *scan;
-
-  if(ptr == NULL) return(0);
-
-  scan = ptr->scan_ptr;
-  for(n = 0; n < ptr->num_scans; n++)
-    FitBSIDScanFreeNext(scan);
-
-  ptr->num_scans = 0;
-  ptr->scan_ptr = (struct FitBSIDScan *)(NULL);
-  ptr->last_ptr = (struct FitBSIDScan *)(NULL);
-  return(1);
-}
-
-void FitBSIDScanFreeNext(struct FitBSIDScan *ptr)
-{
-  int ibm;
-  struct FitBSIDBeam *bm;
-  
-  /* Free the current pointer and cycle to the next one */
-  bm = ptr->bm;
-  for(ibm = 0; ibm < ptr->num_bms; ibm++)
-      FitBSIDBeamFree(&bm[ibm]);
-
-  if(ptr->num_bms > 0)
-    {
-      free(ptr->bm);
-      ptr->num_bms = 0;
-    }
-
-  ptr = ptr->next_scan;
-
-  return;
-}
-
-void FitBSIDBeamFree(struct FitBSIDBeam *ptr)
-{
-  free(ptr->rng);
-  free(ptr->med_rng);
-  free(ptr->front_elv);
-  free(ptr->back_elv);
-  free(ptr->rng_flgs);
-  free(ptr->front_loc);
-  free(ptr->back_loc);
-  return;
-}
