+++ codebase/superdarn/src.bin/tk/tool/make_fov.1.0/test_ut_fov_struct.c
@@ -1,25 +1,15 @@
 /* test_ut_fov_struct.c
    ====================
    Author: Angeline G. Burrell - NRL - 2021
-*/
+   This is a U.S. government work and not under copyright protection in the U.S.
 
-/*
- LICENSE AND DISCLAIMER
+   This file is part of the Radar Software Toolkit (RST).
 
- This file is part of the Radar Software Toolkit (RST).
+   Disclaimer: RST is licensed under GPL v3.0. Please visit 
+               <https://www.gnu.org/licenses/> to see the full license
 
- RST is free software: you can redistribute it and/or modify
- it under the terms of the GNU Lesser General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- any later version.
+   Modifications:
 
- RST is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU Lesser General Public License for more details.
-
- You should have received a copy of the GNU Lesser General Public License
- along with RST.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 #include <stdio.h>
@@ -28,7 +18,7 @@
 
 #include "rtypes.h"
 #include "rtime.h"
-#include "multbsid.h"
+#include "fitmultbsid.h"
 #include "scan_utils.h"
 
 /**
@@ -59,7 +49,7 @@ void test_ut_fov_struct(unsigned char vb, char *vbuf, int nbms, float min_frac,
 			double ut_box_sec, int D_nrg, int E_nrg, int F_nrg,
 			int far_nrg, int D_rgmax, int E_rgmax, int F_rgmax,
 			float D_hmin, float D_hmax, float E_hmax, float F_hmax,
-			struct MultFitBSID *mult_bsid)
+			struct FitMultBSID *mult_bsid)
 {
   int i, iscan, ibm, irg, ireg, ihop, ibox, iwin, istime, ietime, ifov, sbm;
   int bind, bmid, max_pnts, num_bms, max_rg, max_reg, cpid, bmnum, num_rg;
