+++ codebase/superdarn/src.bin/tk/tool/make_fov.1.0/update_scan.c
@@ -1,25 +1,15 @@
 /* update_scan.c
    =============
    Author: Angeline G. Burrell - NRL - 2021
-*/
+   This is a U.S. government work and not under copyright protection in the U.S.
 
-/*
- LICENSE AND DISCLAIMER
+   This file is part of the Radar Software Toolkit (RST).
 
- This file is part of the Radar Software Toolkit (RST).
+   Disclaimer: RST is licensed under GPL v3.0. Please visit 
+               <https://www.gnu.org/licenses/> to see the full license
 
- RST is free software: you can redistribute it and/or modify
- it under the terms of the GNU Lesser General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- any later version.
+   Modifications:
 
- RST is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU Lesser General Public License for more details.
-
- You should have received a copy of the GNU Lesser General Public License
- along with RST.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 #include <stdio.h>
@@ -29,7 +19,7 @@
 
 #include "rmath.h"
 #include "radar.h"
-#include "multbsid.h"
+#include "fitmultbsid.h"
 #include "scan_utils.h"
 
 #ifndef MAX_BMS
@@ -88,7 +78,7 @@ void UpdateScanBSFoV(short int strict_gs, int freq_min, int freq_max,
 		     float D_hmax, float E_hmax, float F_hmax, float D_vh_box,
 		     float E_vh_box, float F_vh_box, float far_vh_box,
 		     float max_hop, struct RadarScan *scan,
-		     struct RadarSite *hard, struct MultFitBSID *mult_bsid)
+		     struct RadarSite *hard, struct FitMultBSID *mult_bsid)
 {
   int iscan, ibm, irg, ifov, ipath, ireg, ivh, igbm, cpid, scan_chan, max_vbin;
   int bind, max_rg, out_num, group_num, bm_num, bmwidth, igood_num;
