+++ codebase/superdarn/src.lib/tk/fitmultbsid.1.0/src/makefile
@@ -1,0 +1,25 @@
+# Makefile for the FitACF MultBSID library
+# ========================================
+# Author: Angeline G. Burrell, NRL (2021)
+# This is a U.S. government work and not under copyright protection in the U.S.
+#
+# Disclaimer: RST is licensed under GPL v3.0. Please visit 
+#             <https://www.gnu.org/licenses/> to see the full license
+#
+# Modifications:
+
+include $(MAKECFG).$(SYSTEM)
+
+INCLUDE = -I$(IPATH)/base -I$(IPATH)/general -I$(IPATH)/superdarn -I../include/
+INC = $(IPATH)/superdarn
+DSTPATH = $(LIBPATH)
+OUTPUT = fitmultbsid
+
+SRC =	fitmultbsid.c \
+	write_ascii.c
+
+OBJS = $(SRC:.c=.o)
+
+LINK = "1"
+
+include $(MAKELIB).$(SYSTEM)
